Add new OAEP_LABEL option for skipping tests
