+++ README.md
@@ -66,7 +66,9 @@ Disabling tests
 To disable specific tests, set the environment variable `CRYPTO11_SKIP=<flags>` where `<flags>` is a comma-separated
 list of the following options:
 
-*  `CERTS` - disables certificate-related tests. Needed for AWS CloudHSM, which doesn't support certificates. 
+*  `CERTS` - disables certificate-related tests. Needed for AWS CloudHSM, which doesn't support certificates.
+*  `OAEP_LABEL` - disables RSA OAEP encryption tests that use source data encoding parameter (also known as a 'label' 
+in some crypto libraries). Needed for AWS CloudHSM.
 
 
 Testing with SoftHSM2
